# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bywise-node",
    "version": "0.0.1",
    "description": "bywise node",
    "contact": {
      "name": "Devel Blockchain",
      "email": "contact@develblockchain.com"
    }
  },
  "paths": {
...
    "/api/v2/transactions/count": {
      "get": {
        "security": [],
        "description": "Count transactions",
        "tags": [
          "TransactionsController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "where",
            "in": "query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TXWhereCount"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "count": {
                      "type": "number"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
    "/api/v2/transactions": {
      "get": {
        "security": [],
        "description": "Get transactions list",
        "tags": [
          "TransactionsController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "where",
            "in": "query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TXWhereQuery"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionWithStatusDTO"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      },
      "post": {
        "security": [],
        "description": "Insert new transactions",
        "tags": [
          "TransactionsController"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
    "/api/v2/transactions/{hash}": {
      "get": {
        "security": [],
        "description": "Get transaction by hash",
        "tags": [
          "TransactionsController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "pattern": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWithStatusDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
    "/api/v2/transactions/fee": {
      "post": {
        "security": [],
        "description": "Calculate transaction fee",
        "tags": [
          "TransactionsController"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "chain": {
                    "required": true,
                    "type": "string"
                  },
                  "from": {
                    "required": true,
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "to": {
                    "required": true,
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "amount": {
                    "required": true,
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "tag": {
                    "type": "string"
                  },
                  "type": {
                    "required": true,
                    "type": "string"
                  },
                  "foreignKeys": {
                    "required": true,
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "data": {
                    "type": "object"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TXWhereCount": {
        "example": "{\n  \"offset\": 0,\n  \"limit\": 100,\n  \"order\": \"desc\",\n  \"status\": \"mined\",\n  \"tx.chain\": \"mainnet\",\n  \"tx.from\": {\n    \"$in\": [\n      \"BWS000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"tx.to\": {\n    \"$in\": [\n      \"BWS000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"tx.tag\": \"\",\n  \"tx.data\": {},\n  \"tx.type\": \"json\",\n  \"tx.foreignKeys\": {\n    \"$in\": [\n      \"key1\"\n    ]\n  },\n  \"slice.hash\": \"string\"\n}",
        "type": "object"
      },
...
      "TXWhereQuery": {
        "example": "{\n  \"offset\": 0,\n  \"limit\": 100,\n  \"order\": \"desc\",\n  \"status\": \"mined\",\n  \"tx.chain\": \"mainnet\",\n  \"tx.from\": {\n    \"$in\": [\n      \"BWS000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"tx.to\": {\n    \"$in\": [\n      \"BWS000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"tx.tag\": \"\",\n  \"tx.data\": {},\n  \"tx.type\": \"json\",\n  \"tx.foreignKeys\": {\n    \"$in\": [\n      \"key1\"\n    ]\n  },\n  \"slice.hash\": \"string\"\n}",
        "type": "object"
      },
      "TransactionOutputDTO": {
        "properties": {
          "cost": {
            "type": "number"
          },
          "size": {
            "type": "number"
          },
          "fee": {
            "type": "string"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "error": {
            "type": "string"
          },
          "output": {
            "type": "object"
          }
        },
        "type": "object"
      },
      "TransactionDTO": {
        "properties": {
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "validator": {
            "type": "string"
          },
          "from": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "amount": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tag": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "foreignKeys": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "data": {
            "type": "object"
          },
          "created": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "validatorSign": {
            "type": "string"
          },
          "sign": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "TxDTO": {
        "properties": {
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "validator": {
            "type": "string"
          },
          "from": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "amount": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tag": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "foreignKeys": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "data": {
            "type": "object"
          },
          "created": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "validatorSign": {
            "type": "string"
          },
          "sign": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TransactionWithStatusDTO": {
        "properties": {
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "validator": {
            "type": "string"
          },
          "from": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "amount": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tag": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "foreignKeys": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "data": {
            "type": "object"
          },
          "created": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "validatorSign": {
            "type": "string"
          },
          "sign": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "jwt": []
    }
  ]
}